# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMONTA BENCANA API Documentation",
        "description": "Sistem Informasi Manajemen Bencana Terpadu - RESTful API untuk pengelolaan data bencana, laporan, monitoring, dan integrasi dengan BMKG.",
        "contact": {
            "email": "support@simonta-bencana.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/bpbd/notifications": {
            "post": {
                "tags": [
                    "BPBD Management"
                ],
                "summary": "Send BPBD Notifications",
                "description": "Endpoint untuk mengirim notifikasi ke publik (Petugas BPBD).",
                "operationId": "bpbdSendNotifications",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "judul",
                                    "pesan",
                                    "target"
                                ],
                                "properties": {
                                    "judul": {
                                        "description": "Judul notifikasi",
                                        "type": "string",
                                        "example": "Peringatan Banjir"
                                    },
                                    "pesan": {
                                        "description": "Isi pesan notifikasi",
                                        "type": "string",
                                        "example": "Warga diminta waspada terhadap potensi banjir"
                                    },
                                    "target": {
                                        "description": "Target penerima notifikasi",
                                        "type": "string",
                                        "enum": [
                                            "Semua",
                                            "Warga",
                                            "Petugas",
                                            "Admin"
                                        ],
                                        "example": "Semua"
                                    },
                                    "kategori": {
                                        "description": "Kategori notifikasi",
                                        "type": "string",
                                        "enum": [
                                            "Info",
                                            "Peringatan",
                                            "Darurat"
                                        ],
                                        "example": "Peringatan"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notification sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "BPBD send notifications"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - BPBD access required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/citizen/evacuation-info": {
            "get": {
                "tags": [
                    "Citizen Access"
                ],
                "summary": "Get Evacuation Information",
                "description": "Endpoint untuk mendapatkan informasi evakuasi (Warga).",
                "operationId": "citizenGetEvacuationInfo",
                "responses": {
                    "200": {
                        "description": "Evacuation information retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Citizen evacuation information"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Citizen access required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get User Dashboard",
                "description": "Endpoint untuk mendapatkan dashboard user berdasarkan role.",
                "operationId": "getDashboard",
                "responses": {
                    "200": {
                        "description": "Dashboard accessed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dashboard accessed successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "nama": {
                                                            "type": "string",
                                                            "example": "Admin Test"
                                                        },
                                                        "username": {
                                                            "type": "string",
                                                            "example": "admintest"
                                                        },
                                                        "role": {
                                                            "type": "string",
                                                            "example": "Admin"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "Admin"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT authentication token. Dapatkan token dengan melakukan login ke endpoint /api/auth/login menggunakan username \"admintest\" dan password \"123456\"."
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}